scripts/requirements.txt
selenium>=4.0.0
pyautogui>=0.9.54
Pillow>=9.0.0
